Remove old cookie banner code
The new cookie banner was deployed 1 week ago (18/06/2019).

We need to remove the old cookie banner code as it is now redundant and could be confusing to other developers.
