+++ app/assets/stylesheets/govuk_publishing_components/components/_cookie-banner.scss
@@ -1,5 +1,4 @@
-$govuk-cookie-banner-background: #d5e8f3;
-$govuk-cookie-banner-background-new: govuk-colour("white");
+$govuk-cookie-banner-background: govuk-colour("white");
 $govuk-cookie-banner-text-green: #00823b;
 
 .js-enabled {
@@ -13,19 +12,7 @@ $govuk-cookie-banner-text-green: #00823b;
   @include govuk-font($size: 16);
   padding: govuk-spacing(2) 0;
   background-color: $govuk-cookie-banner-background;
-}
-
-.gem-c-cookie-banner__message {
-  margin-top: 0;
-  margin-bottom: 0;
-}
-
-.gem-c-cookie-banner__hide-link {
-  display: none;
-
-  .js-enabled & {
-    display: inline-block;
-  }
+  border: 2px solid govuk-colour("black");
 }
 
 // Styles for the new version of the cookie banner
@@ -34,7 +21,7 @@ $govuk-cookie-banner-text-green: #00823b;
   border: 2px solid govuk-colour("black");
 }
 
-.gem-c-cookie-banner--new .gem-c-cookie-banner__message {
+.gem-c-cookie-banner__message {
   display: inline-block;
   margin-bottom: govuk-spacing(1);
   padding-right: govuk-spacing(4);
@@ -134,8 +121,9 @@ $govuk-cookie-banner-text-green: #00823b;
 // Override the styles from govuk_template
 // scss-lint:disable IdSelector
 // sass-lint:disable no-ids
-.gem-c-cookie-banner--new#global-cookie-message {
-  background-color: $govuk-cookie-banner-background-new;
+.gem-c-cookie-banner--new#global-cookie-message,
+.gem-c-cookie-banner#global-cookie-message {
+  background-color: $govuk-cookie-banner-background;
   padding: govuk-spacing(4) 0;
   box-sizing: border-box;
