+++ app/views/govuk_publishing_components/components/_cookie_banner.html.erb
@@ -1,55 +1,45 @@
 <%
-  new_cookie_banner ||= false
   id ||= 'global-cookie-message'
-
-  cookie_banner_class = new_cookie_banner ? "gem-c-cookie-banner--new" : "gem-c-cookie-banner"
-
-  cookie_banner_helper = GovukPublishingComponents::Presenters::CookieBannerHelper.new(local_assigns)
-
-  message = cookie_banner_helper.message
+  message ||= "GOV.UK uses cookies to make the site simpler."
 %>
 
-<div id="<%= id %>" class="<%= cookie_banner_class %>" data-module="cookie-banner" role="region" aria-label="cookie banner">
+<div id="<%= id %>" class="gem-c-cookie-banner" data-module="cookie-banner" role="region" aria-label="cookie banner">
   <div class="gem-c-cookie-banner__wrapper govuk-width-container">
     <p class="gem-c-cookie-banner__message"><%= message %></p>
-    <% if new_cookie_banner %>
-      <div class="gem-c-cookie-banner__buttons">
-        <div class="gem-c-cookie-banner__button gem-c-cookie-banner__button-accept">
-          <%= render "govuk_publishing_components/components/button", {
-            text: "Accept cookies",
-            secondary: true,
-            inline_layout: true,
-            data_attributes: {
-              module: "track-click",
-              "accept-cookies": "true",
-              "track-category": "cookieBanner",
-              "track-action": "Cookie banner accepted"
-            }
-          } %>
-        </div>
-        <div class="gem-c-cookie-banner__button gem-c-cookie-banner__button-settings">
-          <%= render "govuk_publishing_components/components/button", {
-            text: "Cookie settings",
-            href: "/help/cookies",
-            secondary: true,
-            inline_layout: true,
-            data_attributes: {
-              module: "track-click",
-              "track-category": "cookieBanner",
-              "track-action": "Cookie banner settings clicked"
-            }
-          } %>
-        </div>
+    <div class="gem-c-cookie-banner__buttons">
+      <div class="gem-c-cookie-banner__button gem-c-cookie-banner__button-accept">
+        <%= render "govuk_publishing_components/components/button", {
+          text: "Accept cookies",
+          secondary: true,
+          inline_layout: true,
+          data_attributes: {
+            module: "track-click",
+            "accept-cookies": "true",
+            "track-category": "cookieBanner",
+            "track-action": "Cookie banner accepted"
+          }
+        } %>
       </div>
-    <% end %>
+      <div class="gem-c-cookie-banner__button gem-c-cookie-banner__button-settings">
+        <%= render "govuk_publishing_components/components/button", {
+          text: "Cookie settings",
+          href: "/help/cookies",
+          secondary: true,
+          inline_layout: true,
+          data_attributes: {
+            module: "track-click",
+            "track-category": "cookieBanner",
+            "track-action": "Cookie banner settings clicked"
+          }
+        } %>
+      </div>
+    </div>
   </div>
 
-  <% if new_cookie_banner %>
-    <div class="gem-c-cookie-banner__confirmation govuk-width-container" tabindex="-1">
-      <p class="gem-c-cookie-banner__confirmation-message">
-        You&#8217;ve accepted all cookies. You can <a class="govuk-link" href="/help/cookies" data-module="track-click" data-track-category="cookieBanner" data-track-action="Cookie banner settings clicked from confirmation">change your cookie settings</a> at any time.
-      </p>
-      <button class="gem-c-cookie-banner__hide-button" data-hide-cookie-banner="true" data-module="track-click" data-track-category="cookieBanner" data-track-action="Hide cookie banner">Hide</button>
-    </div>
-  <% end %>
+  <div class="gem-c-cookie-banner__confirmation govuk-width-container" tabindex="-1">
+    <p class="gem-c-cookie-banner__confirmation-message">
+      You&#8217;ve accepted all cookies. You can <a class="govuk-link" href="/help/cookies" data-module="track-click" data-track-category="cookieBanner" data-track-action="Cookie banner settings clicked from confirmation">change your cookie settings</a> at any time.
+    </p>
+    <button class="gem-c-cookie-banner__hide-button" data-hide-cookie-banner="true" data-module="track-click" data-track-category="cookieBanner" data-track-action="Hide cookie banner">Hide</button>
+  </div>
 </div>
